Improve the display of parse errors
The username is ambiguous when displaying errors, as the root user
can have multiple files.

Reporting errors with the full context and filename should make it
much easier to work out what is happening when something goes wrong.

This also helps identify places where the username is actually used
for function and not just logging.